Od data intersection test (#8)
* Add Intersection model and OD data

* remove deprecated models

* remove external library rapidcsv from codebase

External libraries should never be committed to the main codebase.

* rename od data files to be more meaningful

* remove deprecated command line parsing from main

* change selectRouteWithMaxFlow to a private function

* rename state variable allOdData to odData

* refactor intersection constructor

Move origin variable to a state variable. All variable should be state variables, when appropriate.

* docs(intersection): remove comments from output function

The output function is not the 'log'.

* docs(intersection): cleanup comments

* Remove else block from intersection ext trans func

External transition function is only triggered when a car enters the intersection.
This else statement is unneccesary.

* comments

* fix model name of intersection to be the origin

The model name (id) is the origin field in the OD data.
The intersection selects the best route by comparing its ID against the origins in the OD data.

* refactor(parking_lot): update docs and variable names

---------

Co-authored-by: braedenkloke <braedenkloke@gmail.com>